+++ PropellerCompiler/Utilities.cpp
@@ -357,7 +357,7 @@ bool GetSymbol(int* pLength)
     }
     else
     {
-        *pLength = strlen(pSymbol);
+        *pLength = (int)strlen(pSymbol);
     }
 
     return true;
